Tried to download manifest images
