+++ cineast-core/src/main/java/org/vitrivr/cineast/core/iiif/presentationapi/v2/models/Rendering.java
@@ -1,5 +1,6 @@
-package org.vitrivr.cineast.core.iiif.presentationapi.v2;
+package org.vitrivr.cineast.core.iiif.presentationapi.v2.models;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
@@ -7,6 +8,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  *   @version 1.0
  *   @created 23.06.21
  */
+@JsonIgnoreProperties(ignoreUnknown = true)
 class Rendering {
 
   @JsonProperty("@id")
