+++ cineast-core/src/main/java/org/vitrivr/cineast/core/iiif/presentationapi/v2/models/Structure.java
@@ -1,5 +1,6 @@
-package org.vitrivr.cineast.core.iiif.presentationapi.v2;
+package org.vitrivr.cineast.core.iiif.presentationapi.v2.models;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
 
@@ -8,6 +9,7 @@ import java.util.List;
  *   @version 1.0
  *   @created 23.06.21
  */
+@JsonIgnoreProperties(ignoreUnknown = true)
 class Structure {
 
   @JsonProperty("@type")
