+++ cineast-core/src/test/java/org/vitrivr/cineast/core/iiif/presentationapi/v2/ManifestTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.Test;
+import org.vitrivr.cineast.core.iiif.presentationapi.v2.models.Manifest;
 
 /**
  * @author singaltanmay
